requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Faker==19.13.0
fire==0.5.0
Flask==2.3.3
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
randomname==0.2.1
six==1.16.0
SQLAlchemy==2.0.23
termcolor==2.3.0
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7

